Update robotmain.proto
Added 10 rays instead of 5 per US sensor. 10 rays are pretty much necessary to reliably hit all the closest blocks. Most rays in the cone skip over the blocks (it's the central 2-3 added to the configuration when you code in the maximum number of rays that really do the job), and what ray hits what block is a weird combination of the block position and ray angle. 
It might be good to attach the ultrasonics lower on the robot.